cleanup, typing & formatting
