+++ .github/workflows/ci-dev.yaml
@@ -28,6 +28,9 @@ jobs:
       - name: Check formatting with ruff
         run: poetry run ruff format --check .
 
+      - name: Check import order
+        run: poetry run ruff check --select I
+
       - name: Check ruff for linting
         run: poetry run ruff check .
