+++ examples/03_move_and_set_ios.py
@@ -1,10 +1,9 @@
 import asyncio
 
 from nova import Nova
-from nova.api import models
 from nova.actions import WriteAction, jnt, ptp
+from nova.api import models
 from nova.types import Pose
-
 
 """
 Example: Move the robot and set I/Os on the path.
