+++ nova/__init__.py
@@ -2,8 +2,8 @@ from nova import actions, api, types
 from nova.core.controller import Controller
 from nova.core.motion_group import MotionGroup
 from nova.core.movement_controller import speed_up as speed_up_movement_controller
-from nova.motion_settings import MotionSettings
 from nova.core.nova import Cell, Nova
+from nova.motion_settings import MotionSettings
 
 __all__ = [
     "Nova",
