+++ nova/core/motion_group.py
@@ -1,13 +1,14 @@
+from typing import Callable
+
 import wandelbots_api_client as wb
 from loguru import logger
 
 from nova.actions import Action, CombinedActions, MovementController, MovementControllerContext
 from nova.core.exceptions import LoadPlanFailed, PlanTrajectoryFailed
-from nova.core.movement_controller import move_forward, motion_group_state_to_motion_state
-from nova.gateway import ApiGateway
-from nova.types import InitialMovementStream, LoadPlanResponse, Pose, MotionState
+from nova.core.movement_controller import motion_group_state_to_motion_state, move_forward
 from nova.core.robot_cell import AbstractRobot
-from typing import Callable
+from nova.gateway import ApiGateway
+from nova.types import InitialMovementStream, LoadPlanResponse, MotionState, Pose
 
 MAX_JOINT_VELOCITY_PREPARE_MOVE = 0.2
 START_LOCATION_OF_MOTION = 0.0
