+++ nova/types/__init__.py
@@ -5,8 +5,8 @@ import wandelbots_api_client as wb
 
 from nova.types.collision_scene import CollisionScene
 from nova.types.pose import Pose
-from nova.types.vector3d import Vector3d
 from nova.types.state import MotionState, RobotState
+from nova.types.vector3d import Vector3d
 
 LoadPlanResponse = wb.models.PlanSuccessfulResponse
 InitialMovementStream = AsyncGenerator[wb.models.StreamMoveResponse, None]
