+++ tests/core/test_io.py
@@ -1,6 +1,8 @@
-from nova.core.io import IOAccess, IOType, IOValueType
-from nova import Nova
 import pytest
+
+from nova import Nova
+from nova.core.io import IOAccess, IOType, IOValueType
+
 # from decouple import config
 
 NOVA_API = "http://172.30.1.41"  # config("NOVA_API")
